+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - SwingEntityTableModel.EntityFilterModelFactory now creates String based filter condition models for Entity based columns.
 ### is.codion.swing.framework.ui
 - EntityTablePanel, table status message now indicates whether the result is limited.
+- EntityTablePanel.INCLUDE_FILTER_PANEL now true by default, INCLUDE_CONDITION_PANEL added.
 
 ## 0.17.19
 ### is.codion.common.db